Rename model files to use snake-case filenames
