+++ src/database/models/index.ts
@@ -1,4 +1,4 @@
 // Export all database models
-export * from './prototype';
-export * from './userModel';
-export * from './workspaceModel';
+export * from './prototype-model';
+export * from './user-model';
+export * from './workspace-model';
